refactor: split mattermost module
